Test changes, rename schoolResponse, and no longer export buildIndex
